+++ packages/sdk/src/dapp/dapp.interface.ts
@@ -5,7 +5,7 @@ import type { AccountAddress } from '../auth/auth.interface';
 export interface Dapps {
   create(command: CreateDappCommand): Promise<Dapp>;
 
-  find(): Promise<Dapp | null>;
+  find(query?: FindOneDappQuery): Promise<Dapp | null>;
 
   findAll(query?: FindDappQuery): Promise<ReadOnlyDapp[]>;
 }
@@ -61,6 +61,10 @@ export interface CreateDappCommand {
 
 export interface DappTelegramBotConfiguration {
   token: string;
+}
+
+export interface FindOneDappQuery {
+  address?: AccountAddress;
 }
 
 export interface FindDappQuery {